Optimised perfromance for VR
Switched off skybox in camera on each scene as this was being drawn behind the current sky boxes, so doubling the rendering overhead of the skys.
Switched off dynamic shadows from the directional lights in each scene.
Switched to single pass rendering in line with Oculus recommendations
Removed a 2nd directional light in the Monet scene
These changes make the frame rate 72fps (ideal target) most of the time in all scenes apart from the hub. In the hub, it takes the FPS from around 25 to 55 average which is much better.